requirements.txt
streamlit
python-docx
docx2pdf
google-generativeai
python-dotenv
protobuf
